nshlib: Move symbol table initialization from nsh_main to nsh_initialize
